+++ probdiffeq/_probdiffeq/priors.py
@@ -11,26 +11,12 @@ __all__ = [
     "prior_ornstein_uhlenbeck_integrated_diffuse",
     "prior_wiener_integrated",
     "prior_wiener_integrated_diffuse",
-    "state_space_model",
 ]
 C = TypeVar("C", bound=Sequence)
 """A type-variable to describe sequences.
 
 Used to type Taylor coefficients, for example.
 """
-
-
-def state_space_model(ssm_fact="dense"):
-    """Construct an implementation of a factorised state-space model."""
-    if ssm_fact == "dense":
-        return ssm_impl.FactSsmImpl.from_dense()
-
-    if ssm_fact == "blockdiag":
-        return ssm_impl.FactSsmImpl.from_blockdiag()
-
-    if ssm_fact == "isotropic":
-        return ssm_impl.FactSsmImpl.from_isotropic()
-    raise ValueError(ssm_fact)
 
 
 def prior_wiener_integrated(
